veritatis/requirements.txt
fastapi==0.121.0
uvicorn==0.38.0
pymilvus==2.6.3
python-dotenv==1.2.1